requirements.txt
python-telegram-bot
nltk
numpy
pandas
sklearn
